cleanup language server database
