+++ crates/language-server/Cargo.toml
@@ -16,7 +16,6 @@ hir = { path = "../hir", package = "fe-hir" }
 language-server-macros = { path = "../language-server-macros", package = "fe-language-server-macros" }
 hir-analysis = { path = "../hir-analysis", package = "fe-hir-analysis" }
 camino = "1.1.4"
-clap = { version = "4.3", features = ["derive"] }
 common = { path = "../common2", package = "fe-common2" }
 anyhow = "1.0.71"
 lsp-types = "0.94.0"
@@ -40,4 +39,3 @@ rust-embed = "8.3.0"
 futures-batch = "0.6.1"
 tracing = "0.1.40"
 tracing-subscriber = "0.3.18"
-tokio-rayon = "2.1.0"
